# Optimize 'social' plugin on cold load