Measuring the build times of 'insiders' docs **without cache** as follows:
`rm -rf .cache/plugin/social && .venv/bin/mkdocs build 2>&1 | grep built`
I get these results after each commit (2 measurements each to make it more reliable):
```
-----before
INFO     -  Documentation built in 8.46 seconds
INFO     -  Documentation built in 8.43 seconds
-----(Optimize social plugin: cache font objects)
INFO     -  Documentation built in 8.40 seconds
INFO     -  Documentation built in 8.42 seconds
-----(Optimize social plugin: reuse resized logo)
INFO     -  Documentation built in 7.78 seconds
INFO     -  Documentation built in 7.79 seconds
-----(Optimize social plugin: cache word width computation)
INFO     -  Documentation built in 7.11 seconds
INFO     -  Documentation built in 7.11 seconds
-----(Optimize social plugin: parallelize image saving)
INFO     -  Documentation built in 5.41 seconds
INFO     -  Documentation built in 5.52 seconds
```
Build times with cache are not affected.