@oprypin 2022-10-29
Sure it's compatible. Always easy to check.
https://docs.python.org/3/library/concurrent.futures.html
>New in version 3.2.