@squidfunk 2022-11-12
Thanks! Commits were merged into #4628.